requirements.txt
langgraph
langgraph-cli
langgraph-sdk
langgraph-checkpoint-sqlite
langsmith
langchain-community
langchain-core
langchain-openai
langchain-anthropic
langchain-experimental
notebook
python-dotenv
langchain-chroma
azure-identity
scikit-learn
openevals
openai
ipython
pyppeteer
langgraph-supervisor
langgraph-cli[inmem]